+++ packages/validator/src/context.ts
@@ -2,13 +2,20 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {Reader} from '@subql/common';
+import {SubstrateProjectManifestVersioned as AvalancheProjectManifestVersioned} from '@subql/common-avalanche';
 import {CosmosProjectManifestVersioned} from '@subql/common-cosmos';
+import {SubstrateProjectManifestVersioned} from '@subql/common-substrate';
+import {TerraProjectManifestVersioned} from '@subql/common-terra';
 import {IPackageJson} from 'package-json-type';
 
 export interface ContextData {
   projectPath: string;
   pkg: IPackageJson;
-  schema?: CosmosProjectManifestVersioned;
+  schema?:
+    | SubstrateProjectManifestVersioned
+    | TerraProjectManifestVersioned
+    | CosmosProjectManifestVersioned
+    | AvalancheProjectManifestVersioned;
 }
 
 export interface Context {