+++ packages/validator/src/rules/rule.ts
@@ -22,11 +22,15 @@ export interface Rule {
   validate(ctx: Context): boolean | Promise<boolean>;
 }
 
+export const deploymentRules: Rule[] = [
+  new RequireValidManifest(),
+  new RequireCustomDsValidation(),
+  new RequireValidRunner(),
+];
+
 export const commonRules: Rule[] = [
   new RequireBuildScript(),
   new RequireCodegenScript(),
   new RequireCliDep(),
-  new RequireValidManifest(),
-  new RequireCustomDsValidation(),
-  new RequireValidRunner(),
+  ...deploymentRules,
 ];