+++ packages/validator/src/validator.ts
@@ -1,8 +1,12 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import {Reader, ReaderFactory, ReaderOptions} from '@subql/common';
+import {getProjectNetwork, NETWORK_FAMILY, Reader, ReaderFactory, ReaderOptions} from '@subql/common';
+import {parseSubstrateProjectManifest as parseAvalancheProjectManifest} from '@subql/common-avalanche';
 import {parseCosmosProjectManifest} from '@subql/common-cosmos';
+import {parseSubstrateProjectManifest} from '@subql/common-substrate';
+import {parseTerraProjectManifest} from '@subql/common-terra';
+
 import {Context} from './context';
 import {Rule, RuleType} from './rules';
 
@@ -16,10 +20,14 @@ export interface Report {
 export class Validator {
   private readonly rules: Rule[] = [];
 
-  static async create(location: string, opts?: ReaderOptions): Promise<Validator> {
-    return new Validator(await ReaderFactory.create(location, opts), location);
+  static async create(location: string, opts?: ReaderOptions, networkFamily?: NETWORK_FAMILY): Promise<Validator> {
+    return new Validator(await ReaderFactory.create(location, opts), location, networkFamily);
   }
-  constructor(private readonly reader: Reader, private readonly location: string) {}
+  constructor(
+    private readonly reader: Reader,
+    private readonly location: string,
+    private readonly networkFamily?: NETWORK_FAMILY
+  ) {}
 
   addRule(...rules: Rule[]): void {
     this.rules.push(...rules);
@@ -41,11 +49,32 @@ export class Validator {
     });
 
     let schema;
-    try {
-      schema = parseCosmosProjectManifest(rawSchema);
-    } catch (e) {
-      console.error(`Load Cosmos project failed, please check the manifest file.`);
-      throw e;
+
+    const networkFamily = this.networkFamily ?? getProjectNetwork(rawSchema);
+    switch (networkFamily) {
+      case NETWORK_FAMILY.substrate:
+        schema = parseSubstrateProjectManifest(rawSchema);
+        if (schema.isV0_0_1) {
+          reports.push({
+            name: 'package-json-file',
+            description: 'A valid `package.json` file must exist in the root directory of the project',
+            valid: !!pkg,
+            skipped: false,
+          });
+        }
+        break;
+      case NETWORK_FAMILY.terra:
+        schema = parseTerraProjectManifest(rawSchema);
+        break;
+      case NETWORK_FAMILY.avalanche:
+        schema = parseAvalancheProjectManifest(rawSchema);
+        break;
+      case NETWORK_FAMILY.cosmos:
+        schema = parseCosmosProjectManifest(rawSchema);
+        break;
+      default:
+        console.error(`Load project failed, please check the manifest file.`);
+        break;
     }
 
     const ctx: Context = {